Issue #3236 - Implement timers (setTimeout/setInterval) for workers
