+++ components/script/dom/webidls/WorkerGlobalScope.webidl
@@ -22,7 +22,7 @@ partial interface WorkerGlobalScope { // not obsolete
   void importScripts(DOMString... urls);
   readonly attribute WorkerNavigator navigator;
 };
-//WorkerGlobalScope implements WindowTimers;
+WorkerGlobalScope implements WindowTimers;
 WorkerGlobalScope implements WindowBase64;
 
 // Proprietary
