+++ components/script/lib.rs
@@ -217,3 +217,4 @@ pub mod html {
 pub mod layout_interface;
 pub mod page;
 pub mod script_task;
+mod timers;
