+++ tests/wpt/metadata/XMLHttpRequest/xmlhttprequest-timeout-worker-twice.html.ini
@@ -1,3 +1,4 @@
 [xmlhttprequest-timeout-worker-twice.html]
   type: testharness
+  disabled: xhr issue #3630
   expected: TIMEOUT
