+++ tests/wpt/metadata/html/dom/interfaces.html.ini
@@ -9333,19 +9333,7 @@
   [WorkerGlobalScope interface: operation createImageBitmap(ImageBitmapSource,long,long,long,long)]
     expected: FAIL
 
-  [WorkerGlobalScope interface: operation setTimeout(Function,long,any)]
-    expected: FAIL
-
-  [WorkerGlobalScope interface: operation setTimeout(DOMString,long,any)]
-    expected: FAIL
-
   [WorkerGlobalScope interface: operation clearTimeout(long)]
-    expected: FAIL
-
-  [WorkerGlobalScope interface: operation setInterval(Function,long,any)]
-    expected: FAIL
-
-  [WorkerGlobalScope interface: operation setInterval(DOMString,long,any)]
     expected: FAIL
 
   [WorkerGlobalScope interface: operation clearInterval(long)]
