--- tests/wpt/metadata/workers/interfaces/WorkerUtils/WindowTimers/001.html.ini
@@ -1,6 +1,0 @@
-[001.html]
-  type: testharness
-  expected: TIMEOUT
-  [setTimeout]
-    expected: TIMEOUT
-
