--- tests/wpt/metadata/workers/interfaces/WorkerUtils/WindowTimers/003.html.ini
@@ -1,6 +1,0 @@
-[003.html]
-  type: testharness
-  expected: TIMEOUT
-  [setInterval]
-    expected: TIMEOUT
-
